Add hack to allow PDINSTANCE-enabled Pds to at least attempt to load externs that use traditional "&s_" style symbols, by catching them in some common functions and aliasing them to the PDINSTANCE-worthy symbols.  To enable this compile libpd with "make MORECFLAGS=-DVST_CLEANSER"
Theoretically this should allow pureVST to load "externs" that are
compiled for inslucsion in Pd vanilla.  (But if you're using Ableton you
will have to adapt the externs to avoid using "class_new",
"class_addmethod", and "post" which conflich with Cycling74 Max function
calls with the same names.  But now who on earth would want to run Pd
inside Ableton... oh, hmm, [sidles over a bit] :)